+++ HISTORY.md
@@ -8,6 +8,7 @@
 
 ### 2020-01-28
 * new dev version
+* gui: record sounds in mono
 
 ## 5.4.4:
 * **Notable Fixes**